Changed the python implementation to be divided to multiple files with the sort in it's own file
