+++ stalinsort.py
@@ -1,27 +1,10 @@
-import random
+# -*- coding: utf-8 -*-
 
-def create_array(array_length, array_range):
-    '''
-    Creates an unsorted array of random length and number range
-    Args:
-        array_length (int): how long the unsorted array will be
-        array_range (int): how wide the range of the numbers in the array will be
-    Returns:
-        list: an unsorted array of integers
-    '''
-
-    unsorted_array = []
-
-    for i in range(1, array_length+1):
-        unsorted_array.append(random.randint(1, array_range))
-
-    return unsorted_array
-
-def stalinSort(unsorted_array):
+def stalin_sort(unsorted_array):
     '''
     Sorts an array, Stalin method
     Args:
-        unsorted_array (list): the unsorted array
+        unsorted_array (list): an array for sorting
     Returns:
         list: the properly sorted array
     '''
@@ -39,14 +22,3 @@ def stalinSort(unsorted_array):
             j += 1
 
     return sorted_array
-
-
-def main():
-    unsorted_array = create_array(100, 100)
-    sorted_array = stalinSort(unsorted_array)
-
-    print("Unsorted: ",unsorted_array)
-    print("Sorted: ",sorted_array)
-
-if __name__ == '__main__':
-    main()
